pyproject.toml
[project]
name = "agno-yaml-builder"
version = "0.1.2"
description = "A Python utility library for defining and constructing Agno Agents, Teams, and Workflows using declarative YAML configuration files"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
license-files = ["LICENSE"]
dependencies = [
    "agno>=1.2.6",
    "duckduckgo-search>=7.5.5",
    "fastapi>=0.115.12",
    "google-genai>=1.8.0",
    "lxml-html-clean>=0.4.1",
    "newspaper4k>=0.9.3.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.55",
]

[project.urls]
Homepage = "https://github.com/zilin/agno-yaml"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["agno_yaml*"]

[project.scripts]
agno-yaml-cli = "argn_yaml.main:main"
